feat(app): use new compilations flags

+++ tests/src/utils.hpp
@@ -3,9 +3,11 @@
 
 #include <QString>
 
+// ===================================================================
+
 namespace Utils {
   inline QString linphoneStringToQString (const std::string &string) {
-    return QString::fromLocal8Bit(string.c_str(), string.size());
+    return QString::fromLocal8Bit(string.c_str(), static_cast<int>(string.size()));
   }
 
   inline std::string qStringToLinphoneString (const QString &string) {
